Update enemy.cpp
mudança de float para int

+++ src/model/enemy.cpp
@@ -1,6 +1,6 @@
 #include "enemy.hpp"
 
-Enemy::Enemy(float pos_x, float pos_y, float speed){
+Enemy::Enemy(float pos_x, int pos_y, float speed){
 	this->pos_x = pos_x;
 	this->pos_y = pos_y;
 	this->speed = speed;
@@ -11,7 +11,7 @@ float Enemy::get_pos_x(){
 	return this->pos_x;
 }
 
-float Enemy::get_pos_y(){
+int Enemy::get_pos_y(){
 	return this->pos_y;
 }
 
@@ -19,7 +19,7 @@ float Enemy::get_speed(){
 	return this->speed;
 }
 
-void Enemy::move(float delta_x, float delta_y){
+void Enemy::move(float delta_x, int delta_y){
 	this->pos_x += delta_x;
 	this->pos_y += delta_y;
 }
